// app/interfaces/DesignInfographics.tsx
  export interface MyInfographic {
    infoId: number;
    fullName: string;
    email: string;
    phoneNumber: string;
    infographicTitle: string;
    infographicDetails: string;
    category: string;
    content: string;
    linkAssets: string;
    keyPoints: string;
    colorTheme: string;
    style: string;
    examples: string;
    additionalNotes: string;
    transactionId: string;
    created: string;
  }